Add function now does not accept duplicates
Duplicates not accepted as this does not make sense in a todo list. Implementation of it is not the most efficient as we iterate through to get the list of tasks and then iterate again to look for duplicates. Does not matter at the moment as it is only one user using a very a simple todo-list.